+++ src/main/java/nl/knaw/dans/avbag/EasyPreprocessAvBag.java
@@ -18,14 +18,14 @@ package nl.knaw.dans.avbag;
 
 import lombok.extern.slf4j.Slf4j;
 import nl.knaw.dans.AbstractCommandLineApp;
+import nl.knaw.dans.avbag.command.ConvertCommand;
 import nl.knaw.dans.avbag.config.EasyPreprocessAvBagConfig;
-import nl.knaw.dans.avbag.core.AVConverter;
 import nl.knaw.dans.avbag.core.PseudoFileSources;
 import nl.knaw.dans.lib.util.CliVersionProvider;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
-import java.nio.file.Path;
+import java.io.IOException;
 
 @Command(name = "easy-preprocess-av-bag",
          mixinStandardHelpOptions = true,
@@ -33,7 +33,6 @@ import java.nio.file.Path;
          description = "Preprocesses a bag exported by easy-fedora-to-bag which contains AV materials")
 @Slf4j
 public class EasyPreprocessAvBag extends AbstractCommandLineApp<EasyPreprocessAvBagConfig> {
-    private EasyPreprocessAvBagConfig config;
 
     public static void main(String[] args) throws Exception {
         new EasyPreprocessAvBag().run(args);
@@ -43,34 +42,17 @@ public class EasyPreprocessAvBag extends AbstractCommandLineApp<EasyPreprocessAv
         return "Preprocesses a bag exported by easy-fedora-to-bag which contains AV materials";
     }
 
-    @CommandLine.Parameters(index = "0",
-                            paramLabel = "INPUT_DIR",
-                            description = "The directory containing the AV dataset.")
-    private Path inputDir;
-
-    @CommandLine.Parameters(index = "1",
-                            paramLabel = "OUTPUT_DIR",
-                            description = "The directory where the converted dataset will be stored.")
-    private Path outputDir;
-
     @Override
     public void configureCommandLine(CommandLine commandLine, EasyPreprocessAvBagConfig config) {
-        this.config = config;
-        log.debug("Configuring command line");
-    }
-
-    @Override
-    public Integer call() {
         try {
-            PseudoFileSources pseudoFileSources = new PseudoFileSources(config.getPseudoFileSources());
-            new AVConverter(inputDir, outputDir, config.getStagingDir(), pseudoFileSources)
-                .convertAll();
-            return 0;
-
+            commandLine.addSubcommand(new ConvertCommand(
+                new PseudoFileSources(config.getPseudoFileSources()),
+                config.getStagingDir()
+            ));
         }
-        catch (Exception e) {
-            log.error("Conversion aborted", e);
-            return 1;
+        catch (IOException e) {
+            throw new RuntimeException(e);
         }
+        log.debug("Configuring command line");
     }
 }
